src/backend/pyproject.toml
[tool.poetry]
name = "Template API"
version = "0.1.0"
description = ""
authors = ["jkalasas <alasasjohnkyle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
loguru = "^0.6.0"
tortoise-orm = "^0.19.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
